% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/containerResizer.R
\name{emphasisCard}
\alias{emphasisCard}
\title{Create an Emphasis Card}
\usage{
emphasisCard(..., bg.color = NULL)
}
\arguments{
\item{...}{content of the card}

\item{bg.color}{content background color}
}
\value{
container with a moving border to denote emphasis
}
\description{
Emphasis container
}
\examples{

r2resize::emphasisCard(
  "sample rpkg.net 1",
  "sample rpkg.net 2 ",
  bg.color = "cyan"
)

}
