% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/containerResizer.R
\name{rezcontCard}
\alias{rezcontCard}
\title{Resizable container content holder}
\usage{
rezcontCard(
  ...,
  width.left = 50,
  width.right = 50,
  width.units = "\%",
  splitter.color = "gray",
  bg.left.color = "yellow",
  left.bg.url = NULL,
  right.bg.url = NULL,
  bg.right.color = "lightgray",
  border.color = NULL,
  position = c("vertical", "horizontal"),
  text.left.color = "white",
  text.right.color = "black"
)
}
\arguments{
\item{...}{content}

\item{width.left}{size of the page font in px e.g. 14px}

\item{width.right}{color of the page font in e.g. darkblue or #006699}

\item{width.units}{boolean. TRUE or FALSE, to add resize toobar to tables}

\item{splitter.color}{splitter color e.g red, black or #333333}

\item{bg.left.color}{left panel color e.g red, black or #333333}

\item{left.bg.url}{left panel background image e.g image1.png or https://..image1.png}

\item{right.bg.url}{right panel background image e.g image1.png or https://..image1.png}

\item{bg.right.color}{right panel color e.g red, black or #333333}

\item{border.color}{color of the container e.g. red or #f5f5f5}

\item{text.left.color}{color of left panel text}

\item{text.right.color}{color of right panel text}
}
\value{
Split screen container
}
\description{
Highly customizable and resizable  container content holder
}
\section{Examples for r2resize}{

More examples and demo pages for are located at this link -
\url{https://rpkg.net/package/r2resize}.
}

\examples{

r2resize::splitCard(
  "sample rpkg.net 1",
  "sample rpkg.net 2 ",
  splitter.color = "transparent",
  bg.right.color = "pink",
  bg.left.color = "gray",
  text.left.color = "green",
  text.right.color = "yellow",
  position = "vertical"
)

}
