% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addJquery.R
\name{add.JQuery}
\alias{add.JQuery}
\title{Include stylesheets and scripts}
\usage{
add.JQuery(version = "3.5.1")
}
\arguments{
\item{version}{jQuery version}
}
\value{
A list of files to be inserted in the header of a page
}
\description{
Use the package scripts and stylesheets in a page
}
\examples{
add.JQuery()
}
