% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launch.R
\name{add.resizer}
\alias{add.resizer}
\title{Create settings for Rmarkdown page}
\usage{
add.resizer(
  theme.color = NULL,
  position = c("top", "bottom", "left", "right"),
  font.size = NULL,
  font.color = NULL,
  tables = TRUE,
  images = TRUE
)
}
\arguments{
\item{theme.color}{theme to use for the page}

\item{position}{primary color for the theme}

\item{font.size}{make header sticky, TRUE or FALSE}

\item{font.color}{set a list group icon}

\item{tables}{add the table}

\item{images}{add the resizer for all images}
}
\value{
styles for themeing the page
}
\description{
Set options for the styling of the Rmarkdown theme to use on the page
}
\section{Demos for sciRmdTheme}{

More examples and demo pages for are located at this link -
\url{https://github.com/oobianom/sciRmdTheme}.
}

\examples{
add.resizer(
   theme.color = NULL,
   position = "top",
   font.size = NULL,
   font.color = NULL,
   tables = TRUE,
   images = TRUE
)

}
