% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setResizer.R
\name{add.resizer}
\alias{add.resizer}
\title{Create settings for Rmarkdown page}
\usage{
add.resizer(
  theme.color = NULL,
  position = c("top", "bottom"),
  font.size = NULL,
  font.color = NULL,
  tables = TRUE,
  images = TRUE,
  line.color = NULL,
  thumb.width = NULL,
  thumb.height = NULL,
  line.width = NULL,
  line.height = NULL,
  dim.units = "px"
)
}
\arguments{
\item{theme.color}{theme to use for the page}

\item{position}{primary color for the theme}

\item{font.size}{size of font in px e.g 14px}

\item{font.color}{set a list group icon}

\item{tables}{add the table}

\item{images}{add the resizer for all images}
}
\value{
Inclusion of mini toolbar for images and tables within a page
}
\description{
Set options for the styling of the Rmarkdown theme to use on the page
}
\section{Demos for r2resize}{

More examples and demo pages for are located at this link -
\url{https://rnetwork.obi.obianom.com/package/r2resize}.
}

\examples{
r2resize::add.resizer(
  theme.color = "blue",
  position = "top",
  font.size = "12px",
  font.color = "brown",
  tables = TRUE,
  images = TRUE,
  line.color = "green",
  line.width = 350
)

}
