% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sliderCards.R
\name{elastiCard}
\alias{elastiCard}
\title{Hover elastic card container}
\usage{
elastiCard(
  ...,
  height.px = NULL,
  width.px = NULL,
  border.color = "white",
  border.width.px = 1,
  active.panel = 1
)
}
\arguments{
\item{...}{list of image containers, see example below}

\item{height.px}{height of container as numeric}

\item{width.px}{width of container as numeric}

\item{border.color}{color of border of each container}

\item{border.width.px}{width of the container border}

\item{active.panel}{the active panel or container e.g. 1}
}
\value{
Multiple cards with hover elastic functionality
}
\description{
Automatic elastic card holder for images or navigation items
}
\section{Examples for r2resize}{

More examples and demo pages are located at this link -
\url{https://rpkg.net/package/r2resize}.
}

\examples{

r2resize::elastiCard(
  item1 = c(
  bg = "https://r2resize.obi.obianom.com/m/image1.jpg",
  icon = "edit",
  title="Sample text 1",
  subtitle="A cool subtitle"),
  item2 = c(
  bg = "https://r2resize.obi.obianom.com/m/image2.jpg",
  title="Sample text 2",
  subtitle="A cool subtitle",
  text.color = "brown"
  )
)

}
