% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/containerResizer.R
\name{empahsisCard}
\alias{empahsisCard}
\title{Emphasize Card}
\usage{
empahsisCard(..., bg.color = NULL, border.color = NULL)
}
\arguments{
\item{...}{content}

\item{bg.color}{background color}

\item{border.color}{border color}
}
\value{
Split screen container
}
\description{
Emphasis container
}
\section{Examples for r2resize}{

More examples and demo pages for are located at this link -
\url{https://rpkg.net/package/r2resize}.
}

\examples{

r2resize::empahsisCard(
  "sample rpkg.net 1",
  "sample rpkg.net 2 "
)

}
