% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/containerResizer.R
\name{empahsisCard}
\alias{empahsisCard}
\title{Create an Emphasis Card}
\usage{
empahsisCard(..., bg.color = NULL)
}
\arguments{
\item{...}{content of the card}

\item{bg.color}{content background color}
}
\value{
container with a moving border to denote emphasis
}
\description{
Emphasis container
}
\section{More examples for r2resize}{

More examples and demo pages are located at this link -
\url{https://r2resize.obi.obianom.com}.
}

\examples{

r2resize::empahsisCard(
  "sample rpkg.net 1",
  "sample rpkg.net 2 ",
  bg.color = "cyan"
)

}
