% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sliderCards.R
\name{flexCard}
\alias{flexCard}
\title{Flexible card container}
\usage{
flexCard(
  ...,
  height.px = NULL,
  width.px = NULL,
  border.color = "white",
  border.width.px = 1,
  active.panel = 1,
  title.color = "black",
  descr.color = "black"
)
}
\arguments{
\item{border.color}{border color of the container e.g. red or #f5f5f5}

\item{left}{content on the left or top}

\item{right}{content on the right or bottom}

\item{splitter.color}{splitter color e.g red, black or #333333}

\item{bg.left.color}{left panel color e.g red, black or #333333}

\item{bg.right.color}{right panel color e.g red, black or #333333}

\item{left.bg.url}{left panel background image e.g image1.png or https://..image1.png}

\item{right.bg.url}{right panel background image e.g image1.png or https://..image1.png}

\item{position}{position of divider or splitter}

\item{text.left.color}{color of left panel text}

\item{text.right.color}{color of right panel text}
}
\value{
Multiple cards with flex functionality
}
\description{
Expandable flexible card holder for images or navigation items
}
\section{Examples for r2resize}{

More examples and demo pages are located at this link -
\url{https://rpkg.net/package/r2resize}.
}

\examples{

r2resize::flexCard(
  "Sample text",
  "Sample text 2"
)

}
