% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r2resize-package.R
\docType{package}
\name{r2resize-package}
\alias{r2resize}
\alias{r2resize-package}
\title{r2resize: In-Text Resize for Images, Tables and Fancy Resize Containers in 'shiny', 'rmarkdown' and 'quarto' Documents}
\description{
Automatic resizing toolbar for containers, images and tables. Various resizer or expandable container functionalities are also included. Most suitable to include in 'shiny', 'markdown' and 'quarto' documents.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://r2resize.obi.obianom.com}
  \item Report bugs at \url{https://github.com/oobianom/r2resize/issues}
}

}
\author{
\strong{Maintainer}: Obinna Obianom \email{idonshayo@gmail.com}

}
\keyword{internal}
