% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/novoImageResizer.R
\name{shinyExpandImage}
\alias{shinyExpandImage}
\title{Add shiny resize for images}
\usage{
shinyExpandImage(imageid)
}
\description{
Elegant image viewer and resizer for images
}
\section{Use case}{

Use in a shiny application for image(s) that you'd like to carry a viewer feature on click.\cr\cr
   - Very portable
}

