% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/containerResizer.R
\name{splitCard}
\alias{splitCard}
\title{Configure toolbar settings for the page}
\usage{
splitCard(
  left,
  right,
  width.left = 50,
  width.right = 50,
  width.units = "\%",
  theme = "dark",
  border = c("none", "primary", "secondary", "warning", "danger", "success"),
  height = NULL,
  position = c("vertical", "horizontal")
)
}
\arguments{
\item{left}{theme color for resizer and table e.g. black or #000000}

\item{right}{position of the resize toolbar e.g. top or bottom}

\item{width.left}{size of the page font in px e.g. 14px}

\item{width.right}{color of the page font in e.g. darkblue or #006699}

\item{width.units}{boolean. TRUE or FALSE, to add resize toobar to tables}

\item{theme}{boolean. TRUE or FALSE, to add resize toobar to images}

\item{border}{color of the resizer track e.g. red or #f5f5f5}

\item{height}{width of the resizer thumb e.g. 25}
}
\value{
Inclusion of mini toolbar for images and tables within a page
}
\description{
Set options for the inclusion of sizing toolbar in documents
}
\section{Examples for r2resize}{

More examples and demo pages for are located at this link -
\url{https://rnetwork.obi.obianom.com/package/r2resize}.
}

\examples{
r2resize::add.resizer() #default settings

r2resize::add.resizer(
  theme.color = "blue",
  position = "top",
  font.size = "12px",
  font.color = "brown",
  tables = TRUE,
  images = TRUE,
  line.color = "green",
  line.width = 150,
  line.height = 5
) # customized settings

}
