% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/containerResizer.R
\name{windowCard}
\alias{windowCard}
\title{Resizable Moveable Expandable Window Card}
\usage{
windowCard(
  ...,
  title = "Sample title",
  width = "50\%",
  bg.color = NULL,
  border.color = NULL,
  header.text.color = NULL,
  body.text.color = NULL
)
}
\arguments{
\item{...}{content of the container}

\item{title}{title of the header}

\item{width}{width of the container}

\item{bg.color}{background color of the content area}

\item{border.color}{color of the container border}

\item{header.text.color}{color of the header text}

\item{body.text.color}{color of the content text}
}
\value{
A window-like card container that is expandable and resizable
}
\description{
Easily expandable and resizable content container holder
}
\note{
Please note that only one windowCard may be created by page
}
\section{Examples for r2resize}{

More examples and demo pages are located at this link -
\url{https://rpkg.net/package/r2resize}.
}

\examples{

\dontrun{
#simple window card with default attributes
#shows up centered on the screen
r2resize::windowCard("sample r2symbols text")


#specify the width of the window card
#specify the title
#specify the header color
#speciy other features
r2resize::windowCard(
title = "Sample Window card",
width = "500px",
bg.color = "white",
border.color = "brown",
header.text.color = "white",
body.text.color = "black")
}

}
